--- a/go.mod
+++ b/go.mod
@@ -35,7 +35,7 @@ require (
 	github.com/hashicorp/vault/api v1.0.4
 	github.com/json-iterator/go v1.1.10
 	github.com/klauspost/compress v1.11.3
-	github.com/klauspost/cpuid v1.3.1
+	github.com/klauspost/cpuid/v2 v2.0.3
 	github.com/klauspost/pgzip v1.2.5
 	github.com/klauspost/readahead v1.3.1
 	github.com/klauspost/reedsolomon v1.9.9
@@ -46,6 +46,7 @@ require (
 	github.com/miekg/dns v1.1.35
 	github.com/minio/cli v1.22.0
 	github.com/minio/highwayhash v1.0.0
+	github.com/minio/md5-simd v1.1.2-0.20210107194844-776275e0c9a7
 	github.com/minio/minio-go/v7 v7.0.7-0.20201217170524-3baf9ea06f7c
 	github.com/minio/selfupdate v0.3.1
 	github.com/minio/sha256-simd v0.1.1