--- a/pkg/s3select/select_test.go
+++ b/pkg/s3select/select_test.go
@@ -22,14 +22,13 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
-	"math"
 	"net/http"
 	"os"
 	"reflect"
 	"strings"
 	"testing"
 
-	"github.com/klauspost/cpuid"
+	"github.com/klauspost/cpuid/v2"
 	"github.com/minio/minio-go/v7"
 	"github.com/minio/simdjson-go"
 )
@@ -362,11 +361,10 @@ func TestJSONQueries(t *testing.T) {
 	for _, testCase := range testTable {
 		t.Run(testCase.name, func(t *testing.T) {
 			// Hack cpuid to the CPU doesn't appear to support AVX2.
-			// Restore whatever happens.
-			defer func(f cpuid.Flags) {
-				cpuid.CPU.Features = f
-			}(cpuid.CPU.Features)
-			cpuid.CPU.Features &= math.MaxUint64 - cpuid.AVX2
+			if cpuid.CPU.Has(cpuid.AVX2) {
+				cpuid.CPU.Disable(cpuid.AVX2)
+				defer cpuid.CPU.Enable(cpuid.AVX2)
+			}
 
 			testReq := testCase.requestXML
 			if len(testReq) == 0 {