## Description

After #9518 make another attempt at using simd-md5 since it now should have less downside.

## How to test this PR?

Before:
```
λ warp put --md5 --benchdata=before -serverprof=cpu -duration=2m
...
Operation: PUT
* Average: 509.67 MiB/s, 50.97 obj/s

Throughput, split into 119 x 1s:
 * Fastest: 915.6MiB/s, 91.56 obj/s
 * 50% Median: 480.2MiB/s, 48.02 obj/s
 * Slowest: 289.9MiB/s, 28.99 obj/s
```

After:
```
λ warp put --md5 --benchdata=after2 -serverprof=cpu -duration=2m 
...
Operation: PUT                                                   
* Average: 458.17 MiB/s, 45.82 obj/s                             
                                                                 
Throughput, split into 119 x 1s:                                 
 * Fastest: 866.5MiB/s, 86.65 obj/s                              
 * 50% Median: 411.2MiB/s, 41.12 obj/s                           
 * Slowest: 244.4MiB/s, 24.44 obj/s                              
````

System is not even close to being CPU limited, but this is too significant a regression to ignore.

## Types of changes
- [x] Optimization (provides speedup with no functional changes)

## Checklist:
- [x] Benchmarks